Rename replay-next/components/elements-new to replay-next/components/elements
