+++ packages/replay-next/components/elements/ElementReactDetails.module.css
