+++ packages/replay-next/components/elements/ElementsList.tsx
@@ -12,9 +12,9 @@ import {
 } from "react";
 import { ListOnItemsRenderedProps } from "react-window";
 
-import { ElementsListData } from "replay-next/components/elements-new/ElementsListData";
-import { NoContentFallback } from "replay-next/components/elements-new/NoContentFallback";
-import { Item } from "replay-next/components/elements-new/types";
+import { ElementsListData } from "replay-next/components/elements/ElementsListData";
+import { NoContentFallback } from "replay-next/components/elements/NoContentFallback";
+import { Item } from "replay-next/components/elements/types";
 import { InlineErrorFallback } from "replay-next/components/errors/InlineErrorFallback";
 import { LoadingProgressBar } from "replay-next/components/LoadingProgressBar";
 import { GenericList } from "replay-next/components/windowing/GenericList";
