+++ packages/replay-next/components/elements/ElementsPanel.module.css
