+++ packages/replay-next/components/elements/ElementsPanel.tsx
@@ -13,9 +13,9 @@ import {
 } from "react";
 import AutoSizer from "react-virtualized-auto-sizer";
 
-import { ElementsList, ImperativeHandle } from "replay-next/components/elements-new/ElementsList";
-import { ElementsListData } from "replay-next/components/elements-new/ElementsListData";
-import { domSearchCache } from "replay-next/components/elements-new/suspense/DOMSearchCache";
+import { ElementsList, ImperativeHandle } from "replay-next/components/elements/ElementsList";
+import { ElementsListData } from "replay-next/components/elements/ElementsListData";
+import { domSearchCache } from "replay-next/components/elements/suspense/DOMSearchCache";
 import Icon from "replay-next/components/Icon";
 import { PanelLoader } from "replay-next/components/PanelLoader";
 import { ReplayClientContext } from "shared/client/ReplayClientContext";
