+++ packages/replay-next/components/elements/hooks/useElementsListItemContextMenu.tsx
@@ -1,9 +1,9 @@
 import { PauseId } from "@replayio/protocol";
 import { ContextMenuItem, useContextMenu } from "use-context-menu";
 
-import { ElementsListData } from "replay-next/components/elements-new/ElementsListData";
-import { elementCache } from "replay-next/components/elements-new/suspense/ElementCache";
-import { Item } from "replay-next/components/elements-new/types";
+import { ElementsListData } from "replay-next/components/elements/ElementsListData";
+import { elementCache } from "replay-next/components/elements/suspense/ElementCache";
+import { Item } from "replay-next/components/elements/types";
 import Icon from "replay-next/components/Icon";
 import { copyToClipboard } from "replay-next/components/sources/utils/clipboard";
 import { ReplayClientInterface } from "shared/client/types";
