+++ packages/replay-next/components/elements/utils/getDistanceFromRoot.ts
